make history collection rather than doc oriented
This assures we will never go over the per document size limit (currently 4MB). Last year in a huurty we ended up with all the schedule data stored with each cme worksheet. This worked out to be about 250K. So in only 16 updates we would have broken the 4MB barrier. While 250K was way too much it worked for our needs at the time.

I also made the history property a function that returns a cursor. This means we can make flexible queries, limit results, select result fields etc. Because it is a cursor we also don't retrive a history item until we need it. Pephaps empty result should return an empty list instead of StopIteration.

This means I added a self.orig attr that stores a copy of the original. This gets updated on update as well.

Gonna add more tests