mod MongoUploadTempStore for modern pymongo/gridfs
It looks as if grids has changed a bunch since this was originally done. It now lets you store your stuff at the top of the document and doesn't generate it's own contentType anymore. So there is no longer a need to store stuff in metadata nor to duplicate info in md.

Also added tests